refactor: 이미지 업로드 에러 문서화
