+++ capturecat-core/src/test/java/com/capturecat/core/api/image/ImageApiTest.java
@@ -4,6 +4,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Stream;
 
+import com.capturecat.core.api.image.dto.AddTagsToImageRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
